Reorganise HEALPix-related functions into internal submodules
