+++ easygems/healpix/utils.py
@@ -1,0 +1,24 @@
+import numpy as np
+
+from .inspection import get_npix
+
+
+def get_full_chunks(indices, chunksize):
+    """Return indices of complete chunks, given a list of indices and a chunksize."""
+    used_chunks = np.unique(np.asarray(indices) // chunksize)
+
+    return (used_chunks[:, np.newaxis] * chunksize + np.arange(chunksize)).flatten()
+
+
+def broadcast_array(dx, fill_value=np.nan):
+    """Broadcast a limited-area HEALPix array to full shape."""
+    arr = np.full_like(dx, fill_value, shape=get_npix(dx))
+    arr[dx.cell] = dx.values
+
+    return arr
+
+
+__all__ = [
+    "get_full_chunks",
+    "broadcast_array",
+]
